+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/internal/options/AnimationOptions.kt
@@ -13,6 +13,6 @@ data class AnimationOptions(
     override var disableNotification: Boolean? = null,
     override var replyParameters: ReplyParameters? = null,
     override var protectContent: Boolean? = null,
-    override var messageThreadId: Long? = null,
+    override var messageThreadId: Int? = null,
     override var hasSpoiler: Boolean? = null,
 ) : OptionsCommon, OptionsParseMode, MediaSpoiler
