+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/internal/options/VideoOptions.kt
@@ -15,6 +15,6 @@ data class VideoOptions(
     override var disableNotification: Boolean? = null,
     override var replyParameters: ReplyParameters? = null,
     override var protectContent: Boolean? = null,
-    override var messageThreadId: Long? = null,
+    override var messageThreadId: Int? = null,
     override var hasSpoiler: Boolean? = null,
 ) : OptionsCommon, OptionsParseMode, FileOptions, MediaSpoiler
